+++ c0321g1_gaming_backend/src/main/java/c0321g1_gaming/model/repository/game/IGameRepository.java
@@ -20,13 +20,13 @@ public interface IGameRepository extends JpaRepository<Game, Long> {
     @Transactional
     @Modifying
     @Query(value = "INSERT INTO game(`name`,content,image,gaming,trailer,game_type_id,flag_delete) VALUES (?1,?2,?3,?4,?5,?6,?7);", nativeQuery = true)
-    void saveGame(String name, String content, String image, int gaming, String trailer, int gameTypeId, int flagDelete);
+    void saveGame(String name, String content, String image, Long gaming, String trailer, int gameTypeId, int flagDelete);
 
     @Modifying
     @Transactional
     @Query(value ="UPDATE game " +
             "SET `name` =?1,content = ?2,image = ?3,gaming=?4,trailer=?5,game_type_id=?6,flag_delete=?7 WHERE game.game_id = ?8" , nativeQuery = true)
-    void updateGame(String name, String content, String image, int gaming, String trailer, int gameTypeId,int flagDelete, Long gameId);
+    void updateGame(String name, String content, String image, Long gaming, String trailer, int gameTypeId,int flagDelete, Long gameId);
 
 
     //    Creator: Thúy
